asaf - added command to find lamprey values by neo value

+++ src/main/java/edu/greenblitz/bigRodika/commands/tests/singleModule/GraphEncoderVoltage.java
@@ -10,7 +10,7 @@ public class GraphEncoderVoltage extends GBCommand {
 
 	private static final long TIMEOUT = 200;
 	private RemoteCSVTarget logger;
-	private LinkedList<Pair<Long, Double>> values;
+	private LinkedList<Pair<Long, Integer>> values;
 	private long t0;
 
 	public GraphEncoderVoltage() {
